Refactor transcribers into separate modules
- Created a new transcribers directory with separate files for each implementation
- Moved the base AudioTranscriber class to its own file
- Created individual files for each transcriber implementation (Whisper API, Whisper MLX, Gemini)
- Added proper imports and improved docstrings
- Updated main.py to use the new module structure

This refactoring improves code organization, makes it easier to maintain
and extend each transcriber type independently, and follows better
separation of concerns.
